pyproject.toml
[project]
name = "pilmoji-for-parser"
version = "0.1.2"
description = "Pilmoji for nonebot-plugin-parser"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiofiles>=25.1.0",
  "emoji>=2.15.0,<3.0.0",
  "httpx>=0.27.2,<1.0.0",
  "pillow>=11.0.0",
]

[project.urls]
IssueTracker = "https://github.com/fllesser/pilmoji-for-parser/issues"
Release = "https://github.com/fllesser/pilmoji-for-parser/releases"
Repository = "https://github.com/fllesser/pilmoji-for-parser"

[dependency-groups]
dev = [
  "ruff>=0.14.7,<1.0.0",
  "poethepoet>=0.38.0,<1.0.0",
  { include-group = "test" },
]
test = [
  "pytest>=9.0.1,<10.0.0",
  "pytest-asyncio>=1.3.0,<2.0.0",
  "pytest-cov>=7.0.0,<8.0.0",
  "pytest-xdist>=3.8.0,<4.0.0",
  "respx>=0.22.0,<1.0.0",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pilmoji"

[tool.bumpversion]
tag = true
commit = true
current_version = "0.1.2"
message = "release: bump vesion from {current_version} to {new_version}"

[[tool.bumpversion.files]]
filename = "uv.lock"
search = "name = \"pilmoji-for-parser\"\nversion = \"{current_version}\""
replace = "name = \"pilmoji-for-parser\"\nversion = \"{new_version}\""

[tool.coverage.report]
show_missing = true # 相当于 --cov-report=term-missing:cite[1]
exclude_lines = [
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@overload",
  "except HTTPError:",
  "assert not",
]

[tool.poe.tasks]
bump = "bump-my-version bump"
show-bump = "bump-my-version show-bump"
test = "pytest --cov=src --cov-report xml --junitxml=./junit.xml -n auto"

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./src",
  ] },
  { root = "./src" },
]
typeCheckingMode = "standard"
disableBytesTypePromotions = true

[tool.pytest]
addopts = [
  "-v",                    # 详细输出
  "-s",                    # 显示打印信息
  "-ra",                   # 显示所有测试结果摘要
  "--tb=short",            # 简短的错误回溯
  "--strict-markers",      # 严格标记模式
  "--import-mode=prepend", # 导入模式
]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "ASYNC", # flake8-async
  "C4",    # flake8-comprehensions
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "PYI",   # flake8-pyi
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
  "TID",   # flake8-tidy-imports
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
  "W191",   # indentation contains tabs
  "TID252", # relative import
]

[tool.ruff.lint.isort]
length-sort = true
known-first-party = ["tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
